Added a new cell to the TeamProfileVC to hold the OVPlotGraphView. Changed the OVPlot yscale to a float, and added a check for numbers > 1000. Implement the segmented control in the OVGraphViewCell

+++ LCSRecap/OVGraphPlotView.h
@@ -11,7 +11,8 @@
 #import "OVGraphYIndicatorView.h"
 @class OVPlotScrollView;
 @interface OVGraphPlotView : UIView <UIScrollViewDelegate>{
-    int yscale;
+    // Changed yscale to a float to support large numbers.
+    float yscale;
     int spacebetweenpoints;
     UIColor *opaquecolor;
     UIColor *lightercolor;
